# openapi.yaml
openapi: "3.1.0"

info:
  title: UIDE Inscripciones API
  version: "2.0.0"
  license:
    name: MIT
    identifier: MIT

servers:
  - url: http://127.0.0.1:3000
    description: Desarrollo local

paths:
  /health:
    get:
      operationId: healthCheck
      summary: Estado del servidor
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Servidor activo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
                ts: "2026-06-19T23:30:00.000Z"

        "401":
          description: API key ausente o inválida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: API key requerida o inválida

  /v1/inscripciones:
    post:
      operationId: crearInscripcionV1
      summary: Crear inscripción (contrato v1)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InscripcionV1"
            example:
              estudianteId: "550e8400-e29b-41d4-a716-446655440000"
              materias:
                - "LTI_05A_458"
              periodoId: "2026-1"

      responses:
        "201":
          description: Inscripción creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InscripcionV1Response"
              example:
                version: v1
                estudianteId: "550e8400-e29b-41d4-a716-446655440000"
                materias:
                  - "LTI_05A_458"
                periodoId: "2026-1"

        "400":
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Campos requeridos: estudianteId, materias, periodoId"

        "401":
          description: API key ausente o inválida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: API key requerida o inválida

  /v2/inscripciones:
    post:
      operationId: crearInscripcionV2
      summary: Crear inscripción con método de pago (contrato v2)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InscripcionV2"
            example:
              estudianteId: "550e8400-e29b-41d4-a716-446655440000"
              materias:
                - "LTI_05A_458"
              periodoId: "2026-1"
              payment_method: scholarship

      responses:
        "201":
          description: Inscripción creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InscripcionV2Response"
              example:
                estudianteId: "550e8400-e29b-41d4-a716-446655440000"
                materias:
                  - "LTI_05A_458"
                periodoId: "2026-1"
                payment_method: scholarship
                version: v2

        "400":
          description: Payload inválido o payment_method fuera del enum
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                campoFaltante:
                  summary: Falta payment_method
                  value:
                    error: "Campos requeridos: estudianteId, materias, periodoId, payment_method"

                metodoInvalido:
                  summary: Método de pago inválido
                  value:
                    error: "payment_method inválido. Valores: debit, credit, scholarship"

        "401":
          description: API key ausente o inválida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: API key requerida o inválida

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key

  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - ts
      properties:
        status:
          type: string
          enum:
            - ok
        ts:
          type: string
          format: date-time
      additionalProperties: false

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
      additionalProperties: false

    InscripcionV1:
      type: object
      required:
        - estudianteId
        - materias
        - periodoId
      properties:
        estudianteId:
          type: string
          format: uuid
        materias:
          type: array
          minItems: 1
          items:
            type: string
        periodoId:
          type: string

    InscripcionV2:
      allOf:
        - $ref: "#/components/schemas/InscripcionV1"
        - type: object
          required:
            - payment_method
          properties:
            payment_method:
              type: string
              enum:
                - debit
                - credit
                - scholarship
      unevaluatedProperties: false

    InscripcionV1Response:
      allOf:
        - $ref: "#/components/schemas/InscripcionV1"
        - type: object
          required:
            - version
          properties:
            version:
              type: string
              enum:
                - v1
      unevaluatedProperties: false

    InscripcionV2Response:
      type: object
      required:
        - estudianteId
        - materias
        - periodoId
        - payment_method
        - version
      properties:
        estudianteId:
          type: string
          format: uuid
        materias:
          type: array
          minItems: 1
          items:
            type: string
        periodoId:
          type: string
        payment_method:
          type: string
          enum:
            - debit
            - credit
            - scholarship
        version:
          type: string
          enum:
            - v2
      additionalProperties: false